Update numpy float call to float32

+++ ml.py
@@ -157,7 +157,7 @@ def feature_importances():
 
     # Load training data
     try:
-        training_data  =np.load(tmp_path+'training_data.npy').astype(np.float)
+        training_data  =np.load(tmp_path+'training_data.npy').astype(np.float32)
         training_labels=np.load(tmp_path+'training_labels.npy')
     except Exception as e:
         print(e)
@@ -252,7 +252,7 @@ def ml_train():
     global NUM_BINS, SAMPLE_RATE
 
     try:
-        training_data=np.load(tmp_path+'training_data.npy').astype(np.float)
+        training_data=np.load(tmp_path+'training_data.npy').astype(np.float32)
     except Exception as e:
         print(e)
         return None, None
@@ -306,7 +306,7 @@ def ml_main():
     
     if is_inferencing:
         try:
-            X_test=np.load(tmp_path+'tmp_frame.npy').astype(np.float)
+            X_test=np.load(tmp_path+'tmp_frame.npy').astype(np.float32)
             assert(X_test.size!=0)
             assert(le is not None)
             assert(model is not None)
